doc-parser-vision-api/requirements.txt
openai==1.35.7
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
fastapi~=0.111.0
uvicorn~=0.30.1
starlette~=0.37.2
pdf2image==1.17.0
pandas
python-doctr==0.9.0
opencv-python==4.10.0.84
torch==2.4.0
torchvision==0.19.0
matplotlib
mindee
pymupdf==1.24.9
google-cloud==0.34.0
google-cloud-vision
poppler-utils==0.1.0
spacy